replace atomic_move with shutil.copyfile in modules replace and lineinfile
atomic_move requires writing access to the folder containing the destination file. But, we only want to copy the source file's content. So, we only need the write permission to the destination file itself and not the directory it is in. If the user has write access only to the file and not to the containing folder, the atomic_move will give a permission error to the user as it is trying to move the source file to the destination folder. But, the shutil.copyfile will successfully copy the source file's contents to the destination file and will keep the stats of the destination file intact.